backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
sentence-transformers
faiss-cpu
beautifulsoup4
python-multipart
